Med Chatbot/requirements.txt
streamlit==1.34.0
langchain==0.1.16
langchain-community>=0.0.32,<0.1
pinecone-client==3.2.2
sentence-transformers==2.6.1
PyPDF2==3.0.1
python-dotenv==1.0.1
tiktoken==0.6.0